+++ front-end/src/App.tsx
@@ -16,7 +16,6 @@ import { ModalProvider } from './context/ModalContext';
 import { NotificationProvider } from './context/NotificationContext';
 import { UserDetailsProvider } from './context/UserDetailsContext';
 import { ApiContextProvider } from './context/ApiContext';
-import AddressLoginForm from './screens/AddressLoginForm';
 import CreatePost from './screens/CreatePost';
 import Discussions from './screens/Discussions';
 import Home from './screens/Home';
@@ -65,9 +64,6 @@ const App = () => {
 													</Route>
 													<Route path="/login">
 														<LoginForm/>
-													</Route>
-													<Route path="/web3login">
-														<AddressLoginForm/>
 													</Route>
 													<Route path="/post/create" >
 														<CreatePost/>
